third_party/WebKit/LayoutTests/fast/multicol/vertical-rl/balancing/balance-trailing-border-after-break.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<script>
    description("Tests that the bottom border at a column boundary in a block that lives in multiple columns is recorded properly for balancing");
</script>
<p>There should be no space between the short black line and the long gray one.</p>
<div id="mc" style="-webkit-writing-mode:vertical-rl; -webkit-columns:3; columns:3; line-height:21px; height:14em; orphans:1; widows:1; border-left:3px solid gray;">
    <div style="width:9px;"></div>
    <div style="border-left:3px solid black;">
        <br><br><br><br><br><br><br><br>
    </div>
</div>
<script>
    shouldBe("mc.offsetWidth", "69");
</script>
